pyproject.toml
[project]
name = "medical-cv-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.1",
    "matplotlib>=3.10.8",
    "monai>=1.5.2",
    "nibabel>=5.4.2",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "pydicom>=3.0.2",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "simpleitk>=2.5.3",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "mypy>=1.20.1",
    "ruff>=0.15.11",
]
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
